+++ SingNSketch/PaletteEditorView.swift
@@ -2,7 +2,7 @@ import Foundation
 import UIKit
 
 class PaletteButton: UIButton {
-    var frequency: Float! = 0
+    var frequency: Double! = 0
     var color: UIColor! = UIColor.blackColor()
 }
